requirements.txt
requests
snowflake-connector-python
flask
pandas